+++ Makefile
@@ -30,13 +30,13 @@ swagger-update:
 	
 swagger-codegen:
 	export GO111MODULE=on
-	go run ./cmd/swagger-codegen/ --output-file ./internal/pkg/handlers/swagger-armspecs.generated.go 
+	go run ./cmd/swagger-codegen/ --output-file ./internal/pkg/expanders/swagger-armspecs.generated.go 
 	# Format the generated code
-	gofmt -s -w internal/pkg/handlers/swagger-armspecs.generated.go
+	gofmt -s -w internal/pkg/expanders/swagger-armspecs.generated.go
 	# Build the generated go files to check for any go build issues
-	go build internal/pkg/handlers/swagger-armspecs.generated.go internal/pkg/handlers/swagger-armspecs.go internal/pkg/handlers/swagger.go internal/pkg/handlers/types.go 
+	go build internal/pkg/expanders/swagger-armspecs.generated.go internal/pkg/expanders/swagger-armspecs.go internal/pkg/expanders/swagger.go internal/pkg/expanders/types.go 
 	# Test the generated code initalizes
-	go test -v internal/pkg/handlers/swagger-armspecs_test.go internal/pkg/handlers/swagger-armspecs.generated.go internal/pkg/handlers/swagger-armspecs.go internal/pkg/handlers/swagger.go internal/pkg/handlers/types.go
+	go test -v internal/pkg/expanders/swagger-armspecs_test.go internal/pkg/expanders/swagger-armspecs.generated.go internal/pkg/expanders/swagger-armspecs.go internal/pkg/expanders/swagger.go internal/pkg/expanders/types.go
 
 debug:
 	GO111MODULE=on go build ./cmd/azbrowse &&  dlv exec ./azbrowse --headless --listen localhost:2345 --api-version 2
