+++ internal/pkg/expanders/appInsights.go
@@ -1,4 +1,4 @@
-package handlers
+package expanders
 
 import (
 	"context"
