+++ internal/pkg/expanders/containerRegistry.go
@@ -1,4 +1,4 @@
-package handlers
+package expanders
 
 import (
 	"bytes"
