+++ internal/pkg/expanders/containerService-swaggerApiSet.go
@@ -1,4 +1,4 @@
-package handlers
+package expanders
 
 import (
 	"bytes"
