+++ internal/pkg/expanders/generic.go
@@ -1,4 +1,4 @@
-package handlers
+package expanders
 
 import (
 	"context"
