+++ internal/pkg/expanders/swagger.go
@@ -1,4 +1,4 @@
-package handlers
+package expanders
 
 import (
 	"context"
